Clarify that Oracle Java is required
This is mentioned a bit further down, but I think marking it clear in both places in beneficial. I didn't notice the first time I checked FAQ (and it worked on my machine with oracle java), but when it failed on my other system with OpenJDK it took me a bit longer to figure out. I hadn't noticed the second comment when reading the first time, just the bold Java 8 note.

So, this adds 'Oracle' to the bold note, so it's harder to miss :)